1/20/1017

+++ SE-4-11/list.cs
@@ -14,8 +14,10 @@ namespace SE_4_11
 {
     public partial class list : Form
     {
-        static MySqlConnection connection = new MySqlConnection("Server = localhost; Database = survey; Uid = root; Password = data;");
+        static string database = "Server = localhost; Database = survey; Uid = root; Password = data;";
+        static MySqlConnection connection = new MySqlConnection(database);
         MySqlCommand command = connection.CreateCommand();
+
         public list()
         {
             InitializeComponent();
@@ -75,13 +77,37 @@ namespace SE_4_11
 
         private void button2_Click(object sender, EventArgs e)
         {
-            int id = Convert.ToInt32(surveys.CurrentRow.Cells["id"].Value);
-            connection.Open();
-            command.CommandText = "DELETE FROM surveys WHERE id = " + id;
-            command.ExecuteNonQuery();
-            connection.Close();
-            MessageBox.Show("Амжилттай устгалаа.");
-            data();
+            if(surveys.CurrentRow.Cells["user_id"].Value == "Надаас")
+            {
+                int id = Convert.ToInt32(surveys.CurrentRow.Cells["id"].Value);
+                command.CommandText = "SELECT id FROM questions WHERE survey_id = " + id;
+                connection.Open();
+                MySqlDataReader reader = command.ExecuteReader();
+                MySqlConnection con = new MySqlConnection(database);
+                MySqlCommand cmd = con.CreateCommand();
+
+                while(reader.Read())
+                {
+                    con.Open();
+                    cmd.CommandText = "DELETE FROM answers WHERE question_id = " + reader["id"];
+                    cmd.ExecuteNonQuery();
+                    cmd.CommandText = "DELETE FROM question_response WHERE question_id = " + reader["id"];
+                    cmd.ExecuteNonQuery();
+                    cmd.CommandText = "DELETE FROM questions WHERE id = " + reader["id"];
+                    cmd.ExecuteNonQuery();
+                    con.Close();
+                }
+
+                reader.Close();
+                command.CommandText = "DELETE FROM surveys WHERE id = " + id;
+                command.ExecuteNonQuery();
+                connection.Close();
+                MessageBox.Show("Амжилттай устгалаа.");
+                data();
+            }else
+            {
+                MessageBox.Show("Уучлаарай, устгаж болохгүй.");
+            }
         }
 
         private void button3_Click(object sender, EventArgs e)
@@ -95,6 +121,20 @@ namespace SE_4_11
         {
             Form1 form = new Form1();
             form.Show();
+        }
+
+        // button for edit
+        private void button5_Click(object sender, EventArgs e)
+        {
+            int surveyId = Convert.ToInt32(surveys.CurrentRow.Cells["id"].Value);
+            Edit editForm = new Edit(surveyId);
+            editForm.Show();
+        }
+
+        // button for update
+        private void button6_Click(object sender, EventArgs e)
+        {
+            data();
         }
     }
 }
